+++ es6module/extensions/contextmenu/jspanel.contextmenu.js
@@ -114,9 +114,3 @@ if (!jsPanel.contextmenu) {
     },false);
 
 }
-
-// Add CommonJS module exports, so it can be imported using require() in Node.js
-// https://nodejs.org/docs/latest/api/modules.html
-if (typeof module !== 'undefined') {
-    module.exports = jsPanel;
-}