+++ es6module/extensions/layout/jspanel.layout.js
@@ -122,9 +122,3 @@ if (!jsPanel.layout) {
     };
 
 }
-
-// Add CommonJS module exports, so it can be imported using require() in Node.js
-// https://nodejs.org/docs/latest/api/modules.html
-if (typeof module !== 'undefined') {
-    module.exports = jsPanel;
-}